+++ src/main/java/com/products/prices/models/Price.java
@@ -20,24 +20,33 @@ import lombok.Setter;
 @NoArgsConstructor
 @Table(name = "PRICES")
 public class Price {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "ID_PRODUCT_PRICE")
     private Long id;
+
     @Column(name = "BRAND_ID", nullable = false)
     private Long brandId;
+
     @Column(name = "START_DATE", nullable = false)
     private Timestamp startDate;
+
     @Column(name = "END_DATE", nullable = false)
     private Timestamp endDate;
+
     @Column(name = "PRICE_LIST", nullable = false)
     private Long priceList;
+
     @Column(name = "PRODUCT_ID", nullable = false)
     private Long productId;
+
     @Column(name = "PRIORITY", nullable = false)
     private Long priority;
+
     @Column(name = "PRICE", nullable = false)
-    private Float priceValue;
+    private Double priceValue;
+
     @Column(name = "CURR", nullable = false)
     private String currency;
